requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
pandas==2.1.1
python-docx==1.1.0
APScheduler==3.10.4
Werkzeug==3.0.1
gunicorn==21.2.0
xlsxwriter==3.1.9
